+++ CHANGELOG.md
@@ -34,6 +34,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+## [v2.5.1] - 2023-12-21
+
+### Bug Fixes
+
+- [707](https://github.com/persistenceOne/pstake-native/pull/707) Fix liquidstakeibc redeem edge case for protecting cValue  
+
 ## [v2.5.0] - 2023-10-20
 
 ### Features